+++ docutils/test/functional/tests/length_units_xetex.py
@@ -1,0 +1,10 @@
+# Source and destination file names
+test_source = "length_units.rst"
+test_destination = "length_units_xetex.tex"
+
+# Keyword parameters passed to publish_file()
+writer = "xetex"
+settings_overrides = {
+    'legacy_column_widths': False,
+    'use_latex_citations': True,
+    }