+++ docutils/test/test_parsers/test_rst/test_directives/test_images.py
@@ -172,15 +172,14 @@ totest['images'] = [
 <document source="test data">
     <image height="100em" uri="picture.png" width="200px">
 """],
-# as in CSS3, units are case-insensitive (new in Docutils 0.22)
 ["""\
 .. image:: picture.png
-   :width:   50 mm
+   :width:  200 Q
    :height: 100 em
 """,
 """\
 <document source="test data">
-    <image height="100em" uri="picture.png" width="50mm">
+    <image height="100em" uri="picture.png" width="200Q">
 """],
 # TODO: support CSS3 units (cf. [feature-requests:#57]
 ["""\
@@ -190,16 +189,7 @@ totest['images'] = [
 """,
 """\
 <document source="test data">
-    <system_message level="3" line="1" source="test data" type="ERROR">
-        <paragraph>
-            Error in "image" directive:
-            invalid option value: (option: "height"; value: '10vh')
-            not a positive number or measure of one of the following units:
-            em, ex, in, cm, mm, pt, pc, px.
-        <literal_block xml:space="preserve">
-            .. image:: picture.png
-               :width: 50%
-               :height: 10vh
+    <image height="10vh" uri="picture.png" width="50%">
 """],
 ["""\
 .. image:: picture.png
@@ -213,7 +203,7 @@ totest['images'] = [
             Error in "image" directive:
             invalid option value: (option: "height"; value: \'40%\')
             not a positive number or measure of one of the following units:
-            em, ex, in, cm, mm, pt, pc, px.
+            em, ex, ch, rem, vw, vh, vmin, vmax, cm, mm, Q, in, pt, pc, px.
         <literal_block xml:space="preserve">
             .. image:: picture.png
                :width: 50%
@@ -230,7 +220,7 @@ totest['images'] = [
             Error in "image" directive:
             invalid option value: (option: "width"; value: \'20mc\')
             not a positive number or measure of one of the following units:
-            em, ex, in, cm, mm, pt, pc, px, %.
+            em, ex, ch, rem, vw, vh, vmin, vmax, cm, mm, Q, in, pt, pc, px, %.
         <literal_block xml:space="preserve">
             .. image:: picture.png
                :width: 20mc